Remove framework module fully
